refactor: extract validation functions and remove duplicates
Extract key validation functions for better testability and code organization:
- Export ValidateDockerConfig, ValidatePath, ValidateShellCommand functions
- Add ExecuteSecureShellChange for secure shell switching
- Create comprehensive validation_test.go with 15+ test cases
- Remove duplicate function declarations causing compilation errors
- Fix security concern in mise.go shell script construction
- Improve code maintainability through module separation
