+++ apps/cli/pkg/commands/setup.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"os"
 	"os/exec"
-	"path/filepath"
-	"regexp"
 	"strings"
 	"time"
 
@@ -16,13 +14,11 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/jameswlane/devex/pkg/config"
-	"github.com/jameswlane/devex/pkg/gitconfig"
 	"github.com/jameswlane/devex/pkg/installers"
 	"github.com/jameswlane/devex/pkg/log"
 	"github.com/jameswlane/devex/pkg/platform"
 	"github.com/jameswlane/devex/pkg/tui"
 	"github.com/jameswlane/devex/pkg/types"
-	"github.com/jameswlane/devex/pkg/utils"
 )
 
 func init() {
@@ -117,22 +113,6 @@ const (
 	PostgreSQLPort = "5432:5432"
 	MySQLPort      = "3306:3306"
 	RedisPort      = "6379:6379"
-)
-
-// Security validation patterns
-var (
-	// validContainerName ensures container names only contain safe characters
-	validContainerName = regexp.MustCompile(`^[a-zA-Z0-9][a-zA-Z0-9_.-]*$`)
-	// validDockerImage ensures image names follow Docker naming conventions
-	validDockerImage = regexp.MustCompile(`^[a-z0-9]([a-z0-9._/-]*[a-z0-9])?:?[a-zA-Z0-9._-]*$`)
-	// validPortMapping ensures port mappings follow expected format
-	validPortMapping = regexp.MustCompile(`^[0-9]+:[0-9]+$`)
-	// validEnvVar ensures environment variables are safe
-	validEnvVar = regexp.MustCompile(`^[A-Z_][A-Z0-9_]*=[a-zA-Z0-9_.-]*$`)
-	// validShellPath ensures shell paths are absolute and contain safe characters
-	validShellPath = regexp.MustCompile(`^/[a-zA-Z0-9/_.-]+$`)
-	// validUsername ensures usernames contain only safe characters
-	validUsername = regexp.MustCompile(`^[a-zA-Z0-9_.-]+$`)
 )
 
 func runGuidedSetup(repo types.Repository, settings config.CrossPlatformSettings) {
@@ -636,28 +616,12 @@ func (m *SetupModel) prevStep() (*SetupModel, tea.Cmd) {
 }
 
 // Helper methods for getting selected items
-func (m *SetupModel) getSelectedShell() string {
-	if m.selectedShell >= 0 && m.selectedShell < len(m.shells) {
-		return m.shells[m.selectedShell]
-	}
-	return m.shells[0] // Default to zsh
-}
 
 func (m *SetupModel) getSelectedLanguages() []string {
 	var selected []string
 	for i, lang := range m.languages {
 		if m.selectedLangs[i] {
 			selected = append(selected, lang)
-		}
-	}
-	return selected
-}
-
-func (m *SetupModel) getSelectedDatabases() []string {
-	var selected []string
-	for i, db := range m.databases {
-		if m.selectedDBs[i] {
-			selected = append(selected, db)
 		}
 	}
 	return selected
@@ -781,491 +745,6 @@ func (m *SetupModel) finalizeSetup(ctx context.Context) error {
 	}
 
 	return nil
-}
-
-func (m *SetupModel) ensureShellInstalled(ctx context.Context, shell string) error {
-	if isToolAvailable(shell) {
-		log.Info("Shell already available", "shell", shell)
-		return nil
-	}
-
-	log.Info("Installing shell", "shell", shell)
-
-	// Get shell app from configuration
-	allApps := m.settings.GetAllApps()
-	for _, app := range allApps {
-		if app.Name == shell {
-			return installers.InstallCrossPlatformApp(app, m.settings, m.repo)
-		}
-	}
-
-	log.Warn("Shell not found in configuration, installing via system package manager", "shell", shell)
-
-	// Fallback to system package manager
-	installCmd := fmt.Sprintf("sudo apt-get update && sudo apt-get install -y %s", shell)
-	output, err := exec.CommandContext(ctx, "bash", "-c", installCmd).CombinedOutput()
-	if err != nil {
-		log.Error("Failed to install shell via apt", err, "shell", shell, "output", string(output))
-		return fmt.Errorf("failed to install %s: %w", shell, err)
-	}
-
-	return nil
-}
-
-func (m *SetupModel) copyShellConfiguration(shell string) error {
-	homeDir := os.Getenv("HOME")
-	devexDir := homeDir + "/.local/share/devex"
-
-	switch shell {
-	case "zsh":
-		return m.copyZshConfiguration(homeDir, devexDir)
-	case "bash":
-		return m.copyBashConfiguration(homeDir, devexDir)
-	case "fish":
-		return m.copyFishConfiguration(homeDir, devexDir)
-	default:
-		log.Warn("No specific configuration available for shell", "shell", shell)
-		return nil
-	}
-}
-
-func (m *SetupModel) copyZshConfiguration(homeDir, devexDir string) error {
-	log.Info("Copying zsh configuration files")
-
-	// Copy main zshrc
-	srcZshrc := devexDir + "/assets/zsh/zshrc"
-	dstZshrc := homeDir + "/.zshrc"
-
-	if err := m.copyFile(srcZshrc, dstZshrc); err != nil {
-		return fmt.Errorf("failed to copy .zshrc: %w", err)
-	}
-
-	// Create a destination directory for zsh config modules
-	zshConfigDir := devexDir + "/defaults/zsh"
-	if err := os.MkdirAll(zshConfigDir, DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create zsh config directory: %w", err)
-	}
-
-	// Copy zsh configuration modules
-	zshFiles := []string{"aliases", "extra", "init", "oh-my-zsh", "prompt", "rc", "shell", "zplug"}
-	for _, file := range zshFiles {
-		src := devexDir + "/assets/zsh/zsh/" + file
-		dst := zshConfigDir + "/" + file
-		if err := m.copyFile(src, dst); err != nil {
-			log.Warn("Failed to copy zsh config file", "file", file, "error", err)
-		}
-	}
-
-	// Copy inputrc
-	inputrcSrc := devexDir + "/assets/zsh/inputrc"
-	inputrcDst := homeDir + "/.inputrc"
-	if err := m.copyFile(inputrcSrc, inputrcDst); err != nil {
-		log.Warn("Failed to copy .inputrc", "error", err)
-	}
-
-	return nil
-}
-
-func (m *SetupModel) copyBashConfiguration(homeDir, devexDir string) error {
-	log.Info("Copying bash configuration files")
-
-	// Copy main bashrc
-	srcBashrc := devexDir + "/assets/bash/bashrc"
-	dstBashrc := homeDir + "/.bashrc"
-
-	if err := m.copyFile(srcBashrc, dstBashrc); err != nil {
-		return fmt.Errorf("failed to copy .bashrc: %w", err)
-	}
-
-	// Create a destination directory for bash config modules
-	bashConfigDir := devexDir + "/defaults/bash"
-	if err := os.MkdirAll(bashConfigDir, DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create bash config directory: %w", err)
-	}
-
-	// Copy bash configuration modules
-	bashFiles := []string{"aliases", "extra", "init", "oh-my-bash", "prompt", "rc", "shell"}
-	for _, file := range bashFiles {
-		src := devexDir + "/assets/bash/bash/" + file
-		dst := bashConfigDir + "/" + file
-		if err := m.copyFile(src, dst); err != nil {
-			log.Warn("Failed to copy bash config file", "file", file, "error", err)
-		}
-	}
-
-	// Copy inputrc
-	inputrcSrc := devexDir + "/assets/bash/inputrc"
-	inputrcDst := homeDir + "/.inputrc"
-	if err := m.copyFile(inputrcSrc, inputrcDst); err != nil {
-		log.Warn("Failed to copy .inputrc", "error", err)
-	}
-
-	// Copy bash_profile if it exists
-	bashProfileSrc := devexDir + "/assets/bash/bash_profile"
-	bashProfileDst := homeDir + "/.bash_profile"
-	if err := m.copyFile(bashProfileSrc, bashProfileDst); err != nil {
-		log.Warn("Failed to copy .bash_profile", "error", err)
-	}
-
-	return nil
-}
-
-func (m *SetupModel) copyFishConfiguration(homeDir, devexDir string) error {
-	log.Info("Copying fish configuration files")
-
-	// Create fish config directory
-	fishConfigDir := homeDir + "/.config/fish"
-	if err := os.MkdirAll(fishConfigDir, DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create fish config directory: %w", err)
-	}
-
-	// Copy main config.fish
-	srcConfig := devexDir + "/assets/fish/config.fish"
-	dstConfig := fishConfigDir + "/config.fish"
-
-	if err := m.copyFile(srcConfig, dstConfig); err != nil {
-		return fmt.Errorf("failed to copy config.fish: %w", err)
-	}
-
-	// Create a destination directory for fish config modules
-	fishDefaultsDir := devexDir + "/defaults/fish"
-	if err := os.MkdirAll(fishDefaultsDir, DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create fish defaults directory: %w", err)
-	}
-
-	// Copy fish configuration modules
-	fishFiles := []string{"aliases", "shell", "init", "prompt"}
-	for _, file := range fishFiles {
-		src := devexDir + "/assets/fish/" + file
-		dst := fishDefaultsDir + "/" + file
-		if err := m.copyFile(src, dst); err != nil {
-			log.Warn("Failed to copy fish config file", "file", file, "error", err)
-		}
-	}
-
-	// Copy fish modules from the fish subdirectory if they exist
-	fishSubFiles := []string{"extra", "oh-my-fish"}
-	for _, file := range fishSubFiles {
-		src := devexDir + "/assets/fish/fish/" + file
-		dst := fishDefaultsDir + "/" + file
-		if err := m.copyFile(src, dst); err != nil {
-			log.Warn("Failed to copy fish config file", "file", file, "error", err)
-		}
-	}
-
-	return nil
-}
-
-func (m *SetupModel) switchToShell(ctx context.Context, shell string) error {
-	shellPath, err := exec.LookPath(shell)
-	if err != nil {
-		return fmt.Errorf("%s not found: %w", shell, err)
-	}
-
-	currentUser := os.Getenv("USER")
-	if currentUser == "" {
-		return fmt.Errorf("unable to determine current user")
-	}
-
-	// Get current shell for comparison
-	currentShell, err := utils.GetUserShell(currentUser)
-	if err != nil {
-		log.Warn("Could not detect current shell", "error", err, "user", currentUser)
-	} else {
-		// Check if the current shell matches the desired shell (compare shell names, not full paths)
-		currentShellName := filepath.Base(currentShell)
-		selectedShellName := filepath.Base(shellPath)
-
-		if currentShellName == selectedShellName {
-			log.Info("User is already using the selected shell", "shell", shell, "currentPath", currentShell, "selectedPath", shellPath, "user", currentUser)
-			m.shellSwitched = false // No switch occurred
-			return nil
-		}
-		log.Info("Current shell differs from selected", "current", currentShell, "selected", shellPath, "user", currentUser)
-	}
-
-	log.Info("Switching to shell", "shell", shell, "path", shellPath, "user", currentUser)
-
-	// Use secure shell change execution with proper validation
-	if err := executeSecureShellChange(ctx, shellPath, currentUser); err != nil {
-		m.shellSwitched = false // Switch failed
-		return err
-	}
-
-	m.shellSwitched = true // Switch succeeded
-	log.Info("Successfully switched shell", "shell", shell)
-	return nil
-}
-
-func (m *SetupModel) copyFile(src, dst string) error {
-	// Validate paths to prevent directory traversal
-	homeDir := os.Getenv("HOME")
-	devexDir := homeDir + "/.local/share/devex"
-
-	// Validate source path is within devex assets directory
-	if err := validatePath(src, devexDir); err != nil {
-		return fmt.Errorf("invalid source path: %w", err)
-	}
-
-	// Validate destination path is within home directory
-	if err := validatePath(dst, homeDir); err != nil {
-		return fmt.Errorf("invalid destination path: %w", err)
-	}
-
-	sourceFile, err := os.Open(src)
-	if err != nil {
-		return fmt.Errorf("failed to open source file %s: %w", src, err)
-	}
-	defer func(sourceFile *os.File) {
-		_ = sourceFile.Close()
-	}(sourceFile)
-
-	destFile, err := os.Create(dst)
-	if err != nil {
-		return fmt.Errorf("failed to create destination file %s: %w", dst, err)
-	}
-	defer func(destFile *os.File) {
-		_ = destFile.Close()
-	}(destFile)
-
-	_, err = sourceFile.WriteTo(destFile)
-	if err != nil {
-		return fmt.Errorf("failed to copy file content: %w", err)
-	}
-
-	return nil
-}
-
-func isToolAvailable(tool string) bool {
-	_, err := exec.LookPath(tool)
-	return err == nil
-}
-
-// copyThemeFiles copies theme assets including backgrounds, neovim colorschemes, and application themes
-func (m *SetupModel) copyThemeFiles() error {
-	log.Info("Copying theme files and configurations")
-
-	homeDir := os.Getenv("HOME")
-	devexDir := homeDir + "/.local/share/devex"
-	assetsDir := devexDir + "/assets"
-
-	// Create necessary directories
-	if err := os.MkdirAll(homeDir+"/.config", DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create .config directory: %w", err)
-	}
-
-	// Copy background images
-	if err := m.copyThemeDirectory(assetsDir+"/themes/backgrounds", homeDir+"/.local/share/backgrounds"); err != nil {
-		log.Warn("Failed to copy background images", "error", err)
-	}
-
-	// Copy Alacritty themes
-	alacrittyConfigDir := homeDir + "/.config/alacritty"
-	if err := os.MkdirAll(alacrittyConfigDir, DirectoryPermissions); err != nil {
-		log.Warn("Failed to create Alacritty config directory", "error", err)
-	} else {
-		if err := m.copyThemeDirectory(assetsDir+"/themes/alacritty", alacrittyConfigDir+"/themes"); err != nil {
-			log.Warn("Failed to copy Alacritty themes", "error", err)
-		}
-	}
-
-	// Copy Neovim colorschemes
-	neovimConfigDir := homeDir + "/.config/nvim"
-	if err := os.MkdirAll(neovimConfigDir+"/colors", DirectoryPermissions); err != nil {
-		log.Warn("Failed to create Neovim config directory", "error", err)
-	} else {
-		if err := m.copyThemeDirectory(assetsDir+"/themes/neovim", neovimConfigDir+"/colors"); err != nil {
-			log.Warn("Failed to copy Neovim colorschemes", "error", err)
-		}
-	}
-
-	// Copy Zellij themes
-	zellijConfigDir := homeDir + "/.config/zellij"
-	if err := os.MkdirAll(zellijConfigDir+"/themes", DirectoryPermissions); err != nil {
-		log.Warn("Failed to create Zellij config directory", "error", err)
-	} else {
-		if err := m.copyThemeDirectory(assetsDir+"/themes/zellij", zellijConfigDir+"/themes"); err != nil {
-			log.Warn("Failed to copy Zellij themes", "error", err)
-		}
-	}
-
-	// Copy Oh My Posh themes
-	ompConfigDir := homeDir + "/.config/oh-my-posh"
-	if err := os.MkdirAll(ompConfigDir+"/themes", DirectoryPermissions); err != nil {
-		log.Warn("Failed to create Oh My Posh config directory", "error", err)
-	} else {
-		if err := m.copyThemeDirectory(assetsDir+"/themes/oh-my-posh", ompConfigDir+"/themes"); err != nil {
-			log.Warn("Failed to copy Oh My Posh themes", "error", err)
-		}
-	}
-
-	// Copy Typora themes
-	typoraThemeDir := homeDir + "/.config/Typora/themes"
-	if err := os.MkdirAll(typoraThemeDir, DirectoryPermissions); err != nil {
-		log.Warn("Failed to create Typora themes directory", "error", err)
-	} else {
-		if err := m.copyThemeDirectory(assetsDir+"/themes/typora", typoraThemeDir); err != nil {
-			log.Warn("Failed to copy Typora themes", "error", err)
-		}
-	}
-
-	// Copy GNOME theme scripts (make them executable)
-	gnomeScriptDir := devexDir + "/themes/gnome"
-	if err := os.MkdirAll(gnomeScriptDir, DirectoryPermissions); err != nil {
-		log.Warn("Failed to create GNOME scripts directory", "error", err)
-	} else {
-		if err := m.copyThemeDirectory(assetsDir+"/themes/gnome", gnomeScriptDir); err != nil {
-			log.Warn("Failed to copy GNOME theme scripts", "error", err)
-		} else {
-			// Make scripts executable
-			m.makeScriptsExecutable(gnomeScriptDir)
-		}
-	}
-
-	log.Info("Theme files copied successfully")
-	return nil
-}
-
-// copyAppConfigFiles copies application configuration files and defaults
-func (m *SetupModel) copyAppConfigFiles() error {
-	log.Info("Copying application configuration files")
-
-	homeDir := os.Getenv("HOME")
-	devexDir := homeDir + "/.local/share/devex"
-	assetsDir := devexDir + "/assets"
-
-	// Create defaults directory in devex
-	defaultsDir := devexDir + "/defaults"
-	if err := os.MkdirAll(defaultsDir, DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create defaults directory: %w", err)
-	}
-
-	// Copy default application configurations
-	if err := m.copyThemeDirectory(assetsDir+"/defaults", defaultsDir); err != nil {
-		log.Warn("Failed to copy default application configurations", "error", err)
-	}
-
-	// Copy XCompose file for special characters
-	xcomposeFile := assetsDir + "/defaults/xcompose"
-	if err := m.copyFile(xcomposeFile, homeDir+"/.XCompose"); err != nil {
-		log.Warn("Failed to copy .XCompose file", "error", err)
-	}
-
-	log.Info("Application configuration files copied successfully")
-	return nil
-}
-
-// setupGitConfiguration applies git configuration using user's name and email
-func (m *SetupModel) setupGitConfiguration() error {
-	log.Info("Setting up git configuration", "name", m.gitFullName, "email", m.gitEmail)
-
-	// Set git user name
-	if m.gitFullName != "" {
-		if _, err := exec.Command("git", "config", "--global", "user.name", m.gitFullName).CombinedOutput(); err != nil {
-			log.Warn("Failed to set git user name", "error", err)
-		} else {
-			log.Info("Git user name set successfully", "name", m.gitFullName)
-		}
-	}
-
-	// Set git user email
-	if m.gitEmail != "" {
-		if _, err := exec.Command("git", "config", "--global", "user.email", m.gitEmail).CombinedOutput(); err != nil {
-			log.Warn("Failed to set git user email", "error", err)
-		} else {
-			log.Info("Git user email set successfully", "email", m.gitEmail)
-		}
-	}
-
-	// Apply additional git configuration from system.yaml
-	homeDir := os.Getenv("HOME")
-	devexDir := homeDir + "/.local/share/devex"
-	systemConfigPath := devexDir + "/config/system.yaml"
-
-	// Load and apply git config if the file exists
-	if gitConfig, err := gitconfig.LoadGitConfig(systemConfigPath); err != nil {
-		log.Warn("Failed to load git configuration from system.yaml", "error", err)
-	} else {
-		if err := gitconfig.ApplyGitConfig(gitConfig); err != nil {
-			log.Warn("Failed to apply git configuration", "error", err)
-		} else {
-			log.Info("Additional git configuration applied successfully")
-		}
-	}
-
-	return nil
-}
-
-// copyThemeDirectory copies all files from source directory to destination directory
-func (m *SetupModel) copyThemeDirectory(srcDir, dstDir string) error {
-	// Validate paths to prevent directory traversal
-	homeDir := os.Getenv("HOME")
-	devexDir := homeDir + "/.local/share/devex"
-
-	// Validate source directory is within devex assets directory
-	if err := validatePath(srcDir, devexDir); err != nil {
-		return fmt.Errorf("invalid source directory: %w", err)
-	}
-
-	// Validate destination directory is within home directory
-	if err := validatePath(dstDir, homeDir); err != nil {
-		return fmt.Errorf("invalid destination directory: %w", err)
-	}
-
-	// Check if source directory exists
-	if _, err := os.Stat(srcDir); os.IsNotExist(err) {
-		return fmt.Errorf("source directory does not exist: %s", srcDir)
-	}
-
-	// Create destination directory
-	if err := os.MkdirAll(dstDir, DirectoryPermissions); err != nil {
-		return fmt.Errorf("failed to create destination directory: %w", err)
-	}
-
-	// Read source directory
-	entries, err := os.ReadDir(srcDir)
-	if err != nil {
-		return fmt.Errorf("failed to read source directory: %w", err)
-	}
-
-	// Copy each file
-	for _, entry := range entries {
-		if !entry.IsDir() {
-			// Validate filename to prevent malicious filenames
-			filename := entry.Name()
-			if strings.Contains(filename, "..") || strings.Contains(filename, "/") || strings.Contains(filename, "\\") {
-				log.Warn("Skipping file with invalid name", "filename", filename)
-				continue
-			}
-
-			srcPath := filepath.Join(srcDir, filename)
-			dstPath := filepath.Join(dstDir, filename)
-
-			if err := m.copyFile(srcPath, dstPath); err != nil {
-				log.Warn("Failed to copy theme file", "src", srcPath, "dst", dstPath, "error", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-// makeScriptsExecutable makes shell scripts executable
-func (m *SetupModel) makeScriptsExecutable(dir string) {
-	entries, err := os.ReadDir(dir)
-	if err != nil {
-		log.Warn("Failed to read directory for making scripts executable", "dir", dir, "error", err)
-		return
-	}
-
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".sh") {
-			scriptPath := filepath.Join(dir, entry.Name())
-			if err := os.Chmod(scriptPath, ExecutablePermissions); err != nil {
-				log.Warn("Failed to make script executable", "script", scriptPath, "error", err)
-			}
-		}
-	}
 }
 
 // Error tracking and validation methods
@@ -1405,89 +884,6 @@ func (m *SetupModel) getLanguageApps() []types.CrossPlatformApp {
 	return apps
 }
 
-// getDockerApp returns a CrossPlatformApp for Docker installation
-func (m *SetupModel) getDockerApp() *types.CrossPlatformApp {
-	return &types.CrossPlatformApp{
-		Name:        "docker",
-		Description: "Container platform for databases and services",
-		Linux: types.OSConfig{
-			InstallMethod:  "apt",
-			InstallCommand: "docker.io",
-		},
-		MacOS: types.OSConfig{
-			InstallMethod:  "brew",
-			InstallCommand: "docker",
-		},
-		Windows: types.OSConfig{
-			InstallMethod:  "winget",
-			InstallCommand: "Docker.DockerDesktop",
-		},
-	}
-}
-
-// getDatabaseApps creates pseudo-apps for database installations via Docker
-func (m *SetupModel) getDatabaseApps() []types.CrossPlatformApp {
-	var apps []types.CrossPlatformApp
-	selectedDBs := m.getSelectedDatabases()
-
-	dbConfigs := map[string]map[string]string{
-		"PostgreSQL": {
-			"image":     "postgres:16",
-			"container": "postgres16",
-			"port":      PostgreSQLPort,
-			"env":       "POSTGRES_HOST_AUTH_METHOD=trust",
-		},
-		"MySQL": {
-			"image":     "mysql:8.4",
-			"container": "mysql8",
-			"port":      MySQLPort,
-			"env":       "MYSQL_ALLOW_EMPTY_PASSWORD=true",
-		},
-		"Redis": {
-			"image":     "redis:7",
-			"container": "redis",
-			"port":      RedisPort,
-			"env":       "",
-		},
-	}
-
-	for _, db := range selectedDBs {
-		if dbConfig, exists := dbConfigs[db]; exists {
-			// Build Docker command securely with validation
-			dockerCmd, err := buildSecureDockerCommand(
-				dbConfig["container"],
-				dbConfig["image"],
-				dbConfig["port"],
-				dbConfig["env"],
-			)
-			if err != nil {
-				log.Error("Invalid Docker configuration for database", err, "database", db)
-				continue // Skip this database if configuration is invalid
-			}
-
-			app := types.CrossPlatformApp{
-				Name:        fmt.Sprintf("docker-%s", strings.ToLower(db)),
-				Description: fmt.Sprintf("Install %s database via Docker", db),
-				Linux: types.OSConfig{
-					InstallMethod:  "docker",
-					InstallCommand: dockerCmd,
-				},
-				MacOS: types.OSConfig{
-					InstallMethod:  "docker",
-					InstallCommand: dockerCmd,
-				},
-				Windows: types.OSConfig{
-					InstallMethod:  "docker",
-					InstallCommand: dockerCmd,
-				},
-			}
-			apps = append(apps, app)
-		}
-	}
-
-	return apps
-}
-
 // getDesktopAppByName finds a desktop app by name from the configuration
 func (m *SetupModel) getDesktopAppByName(name string) *types.CrossPlatformApp {
 	allApps := m.settings.GetAllApps()
@@ -1548,100 +944,6 @@ func printAutomatedSetupPlan() {
 	fmt.Println("  • Programming languages: Node.js, Python")
 	fmt.Println("  • Database: PostgreSQL")
 	fmt.Println()
-}
-
-// validateDockerConfig validates Docker configuration parameters for security
-func validateDockerConfig(containerName, image, portMapping, envVar string) error {
-	if !validContainerName.MatchString(containerName) {
-		return fmt.Errorf("invalid container name: %s", containerName)
-	}
-	if !validDockerImage.MatchString(image) {
-		return fmt.Errorf("invalid Docker image: %s", image)
-	}
-	if !validPortMapping.MatchString(portMapping) {
-		return fmt.Errorf("invalid port mapping: %s", portMapping)
-	}
-	if envVar != "" && !validEnvVar.MatchString(envVar) {
-		return fmt.Errorf("invalid environment variable: %s", envVar)
-	}
-	return nil
-}
-
-// buildSecureDockerCommand constructs a Docker command with proper validation and escaping
-func buildSecureDockerCommand(containerName, image, portMapping, envVar string) (string, error) {
-	// Validate all inputs first
-	if err := validateDockerConfig(containerName, image, portMapping, envVar); err != nil {
-		return "", err
-	}
-
-	// Build command using safe string concatenation (not fmt.Sprintf with user input)
-	var cmdParts []string
-	cmdParts = append(cmdParts, "docker", "run", "-d")
-	cmdParts = append(cmdParts, "--name", containerName)
-	cmdParts = append(cmdParts, "--restart", "unless-stopped")
-	cmdParts = append(cmdParts, "-p", "127.0.0.1:"+portMapping)
-
-	if envVar != "" {
-		cmdParts = append(cmdParts, "-e", envVar)
-	}
-
-	cmdParts = append(cmdParts, image)
-
-	// Join with spaces - this is safe since all parts are validated
-	return strings.Join(cmdParts, " "), nil
-}
-
-// validatePath validates a file path to prevent directory traversal attacks
-func validatePath(path, baseDir string) error {
-	// Clean the path to resolve any .. or . components
-	cleanPath := filepath.Clean(path)
-	cleanBase := filepath.Clean(baseDir)
-
-	// Ensure the path is absolute or convert relative paths
-	if !filepath.IsAbs(cleanPath) {
-		cleanPath = filepath.Join(cleanBase, cleanPath)
-	}
-
-	// Check if the cleaned path is within the base directory
-	relPath, err := filepath.Rel(cleanBase, cleanPath)
-	if err != nil {
-		return fmt.Errorf("invalid path: %w", err)
-	}
-
-	// Ensure the relative path doesn't start with .. (directory traversal)
-	if strings.HasPrefix(relPath, "..") || strings.Contains(relPath, "/../") {
-		return fmt.Errorf("path traversal detected: %s", path)
-	}
-
-	return nil
-}
-
-// validateShellCommand validates parameters for shell command execution
-func validateShellCommand(shellPath, username string) error {
-	if !validShellPath.MatchString(shellPath) {
-		return fmt.Errorf("invalid shell path: %s", shellPath)
-	}
-	if !validUsername.MatchString(username) {
-		return fmt.Errorf("invalid username: %s", username)
-	}
-	return nil
-}
-
-// executeSecureShellChange executes chsh command with proper validation and argument separation
-func executeSecureShellChange(ctx context.Context, shellPath, username string) error {
-	// Validate inputs
-	if err := validateShellCommand(shellPath, username); err != nil {
-		return err
-	}
-
-	// Use exec.Command with separate arguments instead of shell string
-	cmd := exec.CommandContext(ctx, "chsh", "-s", shellPath, username)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		return fmt.Errorf("failed to change shell: %w (output: %s)", err, string(output))
-	}
-
-	return nil
 }
 
 // printAutomatedSetupCompletion displays the completion message and next steps
